*: more forward decl
