+++ dbms/src/Storages/DeltaMerge/ColumnFile/ColumnFileSetReader.h
@@ -15,8 +15,7 @@
 #pragma once
 
 #include <Storages/DeltaMerge/ColumnFile/ColumnFileSetSnapshot.h>
-#include <Storages/DeltaMerge/DMContext.h>
-#include <Storages/DeltaMerge/ScanContext.h>
+#include <Storages/DeltaMerge/DMContext_fwd.h>
 #include <Storages/DeltaMerge/SkippableBlockInputStream.h>
 
 namespace DB
@@ -45,10 +44,7 @@ private:
     LACBytesCollector lac_bytes_collector;
 
 private:
-    explicit ColumnFileSetReader(const DMContext & context_)
-        : context(context_)
-        , lac_bytes_collector(context_.scan_context ? context_.scan_context->resource_group_name : "")
-    {}
+    explicit ColumnFileSetReader(const DMContext & context_);
 
     Block readPKVersion(size_t offset, size_t limit);
