+++ dbms/src/Storages/DeltaMerge/DMContext.h
@@ -16,20 +16,25 @@
 
 #include <Common/Logger.h>
 #include <Core/Types.h>
-#include <Interpreters/Context.h>
+#include <Interpreters/Context_fwd.h>
 #include <Interpreters/Settings.h>
 #include <Storages/DeltaMerge/DMChecksumConfig.h>
 #include <Storages/DeltaMerge/DMContext_fwd.h>
 #include <Storages/DeltaMerge/DeltaMergeDefines.h>
-#include <Storages/DeltaMerge/ScanContext.h>
+#include <Storages/DeltaMerge/ScanContext_fwd.h>
 
 #include <memory>
 
 namespace DB
 {
+
 class StoragePathPool;
 using StoragePathPoolPtr = std::shared_ptr<StoragePathPool>;
 
+class WriteLimiter;
+using WriteLimiterPtr = std::shared_ptr<WriteLimiter>;
+class ReadLimiter;
+using ReadLimiterPtr = std::shared_ptr<ReadLimiter>;
 
 namespace DM
 {
@@ -161,33 +166,7 @@ private:
         size_t rowkey_column_size_,
         const DB::Settings & settings,
         const ScanContextPtr & scan_context_ = nullptr,
-        const String & tracing_id_ = "")
-        : global_context(session_context_.getGlobalContext()) // always save the global context
-        , path_pool(path_pool_)
-        , storage_pool(storage_pool_)
-        , min_version(min_version_)
-        , keyspace_id(keyspace_id_)
-        , physical_table_id(physical_table_id_)
-        , is_common_handle(is_common_handle_)
-        , rowkey_column_size(rowkey_column_size_)
-        , segment_limit_rows(settings.dt_segment_limit_rows)
-        , segment_limit_bytes(settings.dt_segment_limit_size)
-        , segment_force_split_bytes(settings.dt_segment_force_split_size)
-        , delta_limit_rows(settings.dt_segment_delta_limit_rows)
-        , delta_limit_bytes(settings.dt_segment_delta_limit_size)
-        , delta_cache_limit_rows(settings.dt_segment_delta_cache_limit_rows)
-        , delta_cache_limit_bytes(settings.dt_segment_delta_cache_limit_size)
-        , delta_small_column_file_rows(settings.dt_segment_delta_small_column_file_rows)
-        , delta_small_column_file_bytes(settings.dt_segment_delta_small_column_file_size)
-        , stable_pack_rows(settings.dt_segment_stable_pack_rows)
-        , enable_logical_split(settings.dt_enable_logical_split)
-        , read_delta_only(settings.dt_read_delta_only)
-        , read_stable_only(settings.dt_read_stable_only)
-        , enable_relevant_place(settings.dt_enable_relevant_place)
-        , enable_skippable_place(settings.dt_enable_skippable_place)
-        , tracing_id(tracing_id_)
-        , scan_context(scan_context_ ? scan_context_ : std::make_shared<ScanContext>())
-    {}
+        const String & tracing_id_ = "");
 };
 
 } // namespace DM
