+++ dbms/src/Storages/DeltaMerge/File/DMFileBlockInputStream.cpp
@@ -20,6 +20,7 @@
 
 namespace DB::DM
 {
+
 DMFileBlockInputStreamBuilder::DMFileBlockInputStreamBuilder(const Context & context)
     : file_provider(context.getFileProvider())
     , read_limiter(context.getReadLimiter())
@@ -103,4 +104,32 @@ DMFileBlockInputStreamPtr DMFileBlockInputStreamBuilder::build(
 
     return std::make_shared<DMFileBlockInputStream>(std::move(reader), max_sharing_column_count > 0);
 }
+
+DMFileBlockInputStreamPtr createSimpleBlockInputStream(
+    const DB::Context & context,
+    const DMFilePtr & file,
+    ColumnDefines cols)
+{
+    // disable clean read is needed, since we just want to read all data from the file, and we do not know about the column handle
+    // enable read_one_pack_every_time_ is needed to preserve same block structure as the original file
+    DMFileBlockInputStreamBuilder builder(context);
+    if (cols.empty())
+    {
+        // turn into read all columns from file
+        cols = file->getColumnDefines();
+    }
+    return builder.setRowsThreshold(DMFILE_READ_ROWS_THRESHOLD)
+        .onlyReadOnePackEveryTime()
+        .build(file, cols, DB::DM::RowKeyRanges{}, std::make_shared<ScanContext>());
+}
+
+DMFileBlockInputStreamBuilder & DMFileBlockInputStreamBuilder::setFromSettings(const Settings & settings)
+{
+    enable_column_cache = settings.dt_enable_stable_column_cache;
+    max_read_buffer_size = settings.max_read_buffer_size;
+    max_sharing_column_bytes_for_all = settings.dt_max_sharing_column_bytes_for_all;
+    max_sharing_column_count = settings.dt_max_sharing_column_count;
+    return *this;
+}
+
 } // namespace DB::DM
